Initial implementation of syntax injection
