+++ package.json
@@ -21,74 +21,16 @@
     ],
     "icon": "images/icon_128x128.png",
     "contributes": {
-        "languages": [
-            {
-                "id": "markdown",
-                "aliases": [
-                    "Markdown (MyST)"
-                ],
-                "extensions": [
-                    ".myst",
-                    ".myst.md",
-                    ".mnb",
-                    ".mystnb"
-                ],
-                "configuration": "./language-configuration.json"
-            }
-        ],
         "grammars": [
             {
-                "language": "markdown",
-                "scopeName": "text.html.myst",
                 "path": "./syntaxes/myst.tmLanguage",
+                "injectTo": [
+                    "text.html.markdown"
+                ],
+                "scopeName": "text.myst.injection",
                 "embeddedLanguages": {
-                    "meta.embedded.block.css": "css",
-                    "meta.embedded.block.html": "html",
-                    "meta.embedded.block.ini": "ini",
-                    "meta.embedded.block.java": "java",
-                    "meta.embedded.block.lua": "lua",
-                    "meta.embedded.block.makefile": "makefile",
-                    "meta.embedded.block.perl": "perl",
-                    "meta.embedded.block.r": "r",
-                    "meta.embedded.block.ruby": "ruby",
-                    "meta.embedded.block.php": "php",
-                    "meta.embedded.block.sql": "sql",
-                    "meta.embedded.block.vs_net": "vs_net",
-                    "meta.embedded.block.xml": "xml",
-                    "meta.embedded.block.xsl": "xsl",
-                    "meta.embedded.block.yaml": "yaml",
-                    "meta.embedded.block.dosbatch": "dosbatch",
-                    "meta.embedded.block.clojure": "clojure",
-                    "meta.embedded.block.coffee": "coffee",
-                    "meta.embedded.block.c": "c",
-                    "meta.embedded.block.cpp": "cpp",
-                    "meta.embedded.block.diff": "diff",
-                    "meta.embedded.block.dockerfile": "dockerfile",
-                    "meta.embedded.block.go": "go",
-                    "meta.embedded.block.groovy": "groovy",
-                    "meta.embedded.block.pug": "pug",
-                    "meta.embedded.block.javascript": "javascript",
-                    "meta.embedded.block.json": "json",
-                    "meta.embedded.block.jsonc": "jsonc",
-                    "meta.embedded.block.less": "less",
-                    "meta.embedded.block.objc": "objc",
-                    "meta.embedded.block.swift": "swift",
-                    "meta.embedded.block.scss": "scss",
-                    "meta.embedded.block.perl6": "perl6",
-                    "meta.embedded.block.powershell": "powershell",
-                    "meta.embedded.block.python": "python",
-                    "meta.embedded.block.ipython": "python",
-                    "meta.embedded.block.rust": "rust",
-                    "meta.embedded.block.scala": "scala",
-                    "meta.embedded.block.shellscript": "shellscript",
-                    "meta.embedded.block.typescript": "typescript",
-                    "meta.embedded.block.typescriptreact": "typescriptreact",
-                    "meta.embedded.block.csharp": "csharp",
-                    "meta.embedded.block.fsharp": "fsharp",
-                    "meta.embedded.block.dart": "dart",
-                    "meta.embedded.block.handlebars": "handlebars",
-                    "meta.embedded.block.markdown": "markdown",
-                    "meta.embedded.block.log": "log"
+                    "meta.embedded.block.markdown": "text.html.markdown",
+                    "meta.embedded.block.ipython": "source.ipython"
                 }
             }
         ]
@@ -97,7 +39,8 @@
         "js-yaml": "^3.13.1",
         "plist": "^3.0.1",
         "vscode": "^1.1.36",
-        "minimist": ">=1.2.2"
+        "minimist": ">=1.2.2",
+        "nunjucks": "^3.2.1"
     },
     "scripts": {
         "build": "node build",
