+++ template/languages.yaml
@@ -1,0 +1,448 @@
+- identifiers:
+    - css
+    - css.erb
+  language: css
+  name: css
+  sources:
+    - source.css
+- identifiers:
+    - html
+    - htm
+    - shtml
+    - xhtml
+    - inc
+    - tmpl
+    - tpl
+  language: html
+  name: basic
+  sources:
+    - text.html.basic
+- identifiers:
+    - ini
+    - conf
+  language: ini
+  name: ini
+  sources:
+    - source.ini
+- identifiers:
+    - java
+    - bsh
+  language: java
+  name: java
+  sources:
+    - source.java
+- identifiers:
+    - lua
+  language: lua
+  name: lua
+  sources:
+    - source.lua
+- identifiers:
+    - Makefile
+    - makefile
+    - GNUmakefile
+    - OCamlMakefile
+  language: makefile
+  name: makefile
+  sources:
+    - source.makefile
+- identifiers:
+    - perl
+    - pl
+    - pm
+    - pod
+    - t
+    - PL
+    - psgi
+    - vcl
+  language: perl
+  name: perl
+  sources:
+    - source.perl
+- identifiers:
+    - R
+    - r
+    - s
+    - S
+    - Rprofile
+    - \{\.r.+?\}
+  language: r
+  name: r
+  sources:
+    - source.r
+- identifiers:
+    - ruby
+    - rb
+    - rbx
+    - rjs
+    - Rakefile
+    - rake
+    - cgi
+    - fcgi
+    - gemspec
+    - irbrc
+    - Capfile
+    - ru
+    - prawn
+    - Cheffile
+    - Gemfile
+    - Guardfile
+    - Hobofile
+    - Vagrantfile
+    - Appraisals
+    - Rantfile
+    - Berksfile
+    - Berksfile.lock
+    - Thorfile
+    - Puppetfile
+  language: ruby
+  name: ruby
+  sources:
+    - source.ruby
+- identifiers:
+    - php
+    - php3
+    - php4
+    - php5
+    - phpt
+    - phtml
+    - aw
+    - ctp
+  language: php
+  name: php
+  sources:
+    - text.html.basic
+    - source.php
+- identifiers:
+    - sql
+    - ddl
+    - dml
+  language: sql
+  name: sql
+  sources:
+    - source.sql
+- identifiers:
+    - vb
+  language: vs_net
+  name: vs_net
+  sources:
+    - source.asp.vb.net
+- identifiers:
+    - xml
+    - xsd
+    - tld
+    - jsp
+    - pt
+    - cpt
+    - dtml
+    - rss
+    - opml
+  language: xml
+  name: xml
+  sources:
+    - text.xml
+- identifiers:
+    - xsl
+    - xslt
+  language: xsl
+  name: xsl
+  sources:
+    - text.xml.xsl
+- identifiers:
+    - yaml
+    - yml
+  language: yaml
+  name: yaml
+  sources:
+    - source.yaml
+- identifiers:
+    - bat
+    - batch
+  language: dosbatch
+  name: dosbatch
+  sources:
+    - source.batchfile
+- identifiers:
+    - clj
+    - cljs
+    - clojure
+  language: clojure
+  name: clojure
+  sources:
+    - source.clojure
+- identifiers:
+    - coffee
+    - Cakefile
+    - coffee.erb
+  language: coffee
+  name: coffee
+  sources:
+    - source.coffee
+- identifiers:
+    - c
+    - h
+  language: c
+  name: c
+  sources:
+    - source.c
+- alias_name: source.cpp
+  identifiers:
+    - cpp
+    - c\+\+
+    - cxx
+  language: cpp
+  name: cpp
+  sources:
+    - source.cpp
+- identifiers:
+    - patch
+    - diff
+    - rej
+  language: diff
+  name: diff
+  sources:
+    - source.diff
+- identifiers:
+    - dockerfile
+    - Dockerfile
+  language: dockerfile
+  name: dockerfile
+  sources:
+    - source.dockerfile
+- identifiers:
+    - COMMIT_EDITMSG
+    - MERGE_MSG
+  name: git_commit
+  sources:
+    - text.git-commit
+- identifiers:
+    - git-rebase-todo
+  name: git_rebase
+  sources:
+    - text.git-rebase
+- identifiers:
+    - go
+    - golang
+  language: go
+  name: go
+  sources:
+    - source.go
+- identifiers:
+    - groovy
+    - gvy
+  language: groovy
+  name: groovy
+  sources:
+    - source.groovy
+- identifiers:
+    - jade
+    - pug
+  language: pug
+  name: pug
+  sources:
+    - text.pug
+- identifiers:
+    - js
+    - jsx
+    - javascript
+    - es6
+    - mjs
+    - cjs
+    - \{\.js.+?\}
+  language: javascript
+  name: js
+  sources:
+    - source.js
+- identifiers:
+    - regexp
+  name: js_regexp
+  sources:
+    - source.js.regexp
+- identifiers:
+    - json
+    - json5
+    - sublime-settings
+    - sublime-menu
+    - sublime-keymap
+    - sublime-mousemap
+    - sublime-theme
+    - sublime-build
+    - sublime-project
+    - sublime-completions
+  language: json
+  name: json
+  sources:
+    - source.json
+- identifiers:
+    - jsonc
+  language: jsonc
+  name: jsonc
+  sources:
+    - source.json.comments
+- identifiers:
+    - less
+  language: less
+  name: less
+  sources:
+    - source.css.less
+- identifiers:
+    - objectivec
+    - objective-c
+    - mm
+    - objc
+    - obj-c
+    - m
+    - h
+  language: objc
+  name: objc
+  sources:
+    - source.objc
+- identifiers:
+    - swift
+  language: swift
+  name: swift
+  sources:
+    - source.swift
+- identifiers:
+    - scss
+  language: scss
+  name: scss
+  sources:
+    - source.css.scss
+- identifiers:
+    - perl6
+    - p6
+    - pl6
+    - pm6
+    - nqp
+  language: perl6
+  name: perl6
+  sources:
+    - source.perl.6
+- identifiers:
+    - powershell
+    - ps1
+    - psm1
+    - psd1
+  language: powershell
+  name: powershell
+  sources:
+    - source.powershell
+- identifiers:
+    - python
+    - py
+    - py3
+    - rpy
+    - pyw
+    - cpy
+    - SConstruct
+    - Sconstruct
+    - sconstruct
+    - SConscript
+    - gyp
+    - gypi
+    - \{\.python.+?\}
+  language: python
+  name: python
+  sources:
+    - source.python
+- identifiers:
+    - ipython
+    - ipython2
+    - ipython3
+    - \{\.ipython.+?\}
+  language: python
+  name: ipython
+  sources:
+    - source.python
+- identifiers:
+    - re
+  name: regexp_python
+  sources:
+    - source.regexp.python
+- identifiers:
+    - rust
+    - rs
+    - \{\.rust.+?\}
+  language: rust
+  name: rust
+  sources:
+    - source.rust
+- identifiers:
+    - scala
+    - sbt
+  language: scala
+  name: scala
+  sources:
+    - source.scala
+- identifiers:
+    - shell
+    - sh
+    - bash
+    - zsh
+    - bashrc
+    - bash_profile
+    - bash_login
+    - profile
+    - bash_logout
+    - .textmate_init
+    - \{\.bash.+?\}
+  language: shellscript
+  name: shell
+  sources:
+    - source.shell
+- identifiers:
+    - typescript
+    - ts
+  language: typescript
+  name: ts
+  sources:
+    - source.ts
+- identifiers:
+    - tsx
+  language: typescriptreact
+  name: tsx
+  sources:
+    - source.tsx
+- identifiers:
+    - cs
+    - csharp
+    - c#
+  language: csharp
+  name: csharp
+  sources:
+    - source.cs
+- identifiers:
+    - fs
+    - fsharp
+    - f#
+  language: fsharp
+  name: fsharp
+  sources:
+    - source.fsharp
+- identifiers:
+    - dart
+  language: dart
+  name: dart
+  sources:
+    - source.dart
+- identifiers:
+    - handlebars
+    - hbs
+  language: handlebars
+  name: handlebars
+  sources:
+    - text.html.handlebars
+- identifiers:
+    - markdown
+    - md
+  language: markdown
+  name: markdown
+  sources:
+    - text.html.markdown
+- identifiers:
+    - log
+  language: log
+  name: log
+  sources:
+    - text.log
